# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IAEDU API",
        "description": "API para el sistema educativo IAEDU",
        "contact": {
            "email": "admin@iaedu.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Obtener lista de estudiantes",
                "operationId": "6a6bb2a44ed676a073aaa78abc6ac0a4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por nombre o matrícula",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "description": "Filtrar por grupo",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "Incluir relaciones (user,group,grades,attendances)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de estudiantes obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Students retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Student"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Crear nuevo estudiante",
                "operationId": "1b3c5a8823e6d516e0cd009031603333",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "matricula",
                                    "nombre",
                                    "apellido_paterno",
                                    "apellido_materno",
                                    "group_id"
                                ],
                                "properties": {
                                    "matricula": {
                                        "type": "string",
                                        "example": "2024001"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido_paterno": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "apellido_materno": {
                                        "type": "string",
                                        "example": "García"
                                    },
                                    "group_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2005-03-15"
                                    },
                                    "curp": {
                                        "type": "string",
                                        "example": "PEGJ050315HDFXXX01"
                                    },
                                    "blood_type": {
                                        "type": "string",
                                        "example": "O+"
                                    },
                                    "allergies": {
                                        "type": "string",
                                        "example": "Polen"
                                    },
                                    "emergency_contact": {
                                        "type": "object"
                                    },
                                    "parent_data": {
                                        "type": "object"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Estudiante creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Student created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Student"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
...
        "/api/students/{id}/grades": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Obtener calificaciones del estudiante",
                "operationId": "4ebb3f5768960c71aac2303c406f49ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del estudiante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calificaciones obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Student grades retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Grade"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estudiante no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}